refactor(cli): remove legacy nightshift.go file
Phase 6.B of #131. The filename was historical — there's no
`belayer nightshift` subcommand. Split contents into:
  - spawn.go             (newSpawnCmd, spawnAgentRequest, SpawnAgent client)
  - roster.go            (newRosterCmd, AgentRunView, ListAgents client)
  - agent_lifecycle.go   (newFinishCmd, newRequestCompletionCmd, FinishAgent client)

No behavior changes. Test renamed from TestRootCmdRegistersNightshiftCommands
to TestRootCmdRegistersAgentCommands.

Refs #131, #144.
